jos td !== null

+++ src/main/java/toplana/web/rest/StavkeIzvodaResource.java
@@ -130,7 +130,7 @@ public class StavkeIzvodaResource {
         	stavkeIzvoda.setRasporedjena(true);        	
         	Transakcija td = stavkeIzvodaTransakcija.getStavkeIzvoda().getTransakcija();
         	
-        	if (td.getId() != null && td.getId() > 0) {
+        	if (td != null && td.getId() != null && td.getId() > 0) {
         		
         		transakcijaRepository.delete(td);
         		
@@ -157,7 +157,7 @@ public class StavkeIzvodaResource {
                 	
         			Transakcija td = stavkeIzvodaTransakcija.getStavkeIzvoda().getTransakcija();
                 	
-                	if (td.getId() != null && td.getId() > 0) {
+                	if (td != null && td.getId() != null && td.getId() > 0) {
                 		
                 		transakcijaRepository.delete(td);
                 		
@@ -183,7 +183,7 @@ public class StavkeIzvodaResource {
 	            	
 	        		Transakcija td = stavkeIzvodaTransakcija.getStavkeIzvoda().getTransakcija();
 	            	
-	            	if (td.getId() != null && td.getId() > 0) {
+	            	if (td != null && td.getId() != null && td.getId() > 0) {
 	            		
 	            		transakcijaRepository.delete(td);
